feat-test: Exercise 4 - Testing the plant removed

+++ src/main/java/br/com/ero/tests/swplanetapi/domain/PlanetService.java
@@ -1,5 +1,6 @@
 package br.com.ero.tests.swplanetapi.domain;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 
@@ -9,30 +10,33 @@ import java.util.Optional;
 @Service
 public class PlanetService {
 
-    private PlanetRepository planetRepository;
+  private PlanetRepository planetRepository;
 
-    public PlanetService(PlanetRepository planetRepository) {
-        this.planetRepository = planetRepository;
-    }
+  public PlanetService(PlanetRepository planetRepository) {
+    this.planetRepository = planetRepository;
+  }
 
-    public Planet create(Planet planet) {
-        return planetRepository.save(planet);
-    }
+  public Planet create(Planet planet) {
+    return planetRepository.save(planet);
+  }
 
-    public Optional<Planet> get(Long id) {
-        return planetRepository.findById(id);
-    }
+  public Optional<Planet> get(Long id) {
+    return planetRepository.findById(id);
+  }
 
-    public Optional<Planet> getByName(String name) {
-        return planetRepository.findByName(name);
-    }
+  public Optional<Planet> getByName(String name) {
+    return planetRepository.findByName(name);
+  }
 
-    public List<Planet> list(String terrain, String climate) {
-        Example<Planet> query = QueryBuilder.makeQuery(new Planet(climate, terrain));
-        return planetRepository.findAll(query);
-    }
+  public List<Planet> list(String terrain, String climate) {
+    Example<Planet> query = QueryBuilder.makeQuery(new Planet(climate, terrain));
+    return planetRepository.findAll(query);
+  }
 
-    public void remove(Long id) {
-        planetRepository.deleteById(id);
+  public void remove(Long id) {
+    if (planetRepository.findById(id).isPresent()) {
+      planetRepository.deleteById(id);
     }
+    throw new EmptyResultDataAccessException(1);
+  }
 }
